Add StartPage stories and extract component from pages/index.js
Moves the StartPage component into src/StartPage.js so it can be
imported in Storybook without pulling in the prisma server dependency.
Adds three stories: with leaderboard, without leaderboard (empty state),
and no active competition.
