+++ pages/index.js
@@ -1,167 +1,7 @@
-import { format } from 'date-fns';
-import Head from 'next/head';
-import Link from 'next/link';
-import React from 'react';
-
-import Leaderboard from '../src/Leaderboard.js';
-import Menu from '../src/Menu';
-import competitionDateString from '../src/competitionDateString';
-import ensureDates from '../src/ensureDates.js';
 import prisma from '../src/prisma';
+import StartPage from '../src/StartPage.js';
 
-const youtubers = [
-  {
-    name: 'Stefan Idstam',
-    url: 'https://www.youtube.com/@stefanidstamgolf',
-    avatar: {
-      url: 'https://yt3.googleusercontent.com/5qBdv_1t7oHY1rD7rQzH7tq0NMb4OO9OwsOLL99QOMt3cr9M2nZ7-NDOUhuUPPT3MPxKtNecjg=s320-c-k-c0x00ffffff-no-rj',
-      width: 320,
-      height: 320,
-    },
-  },
-  {
-    name: 'Adam Andersson',
-    url: 'https://adamanderssongolf.com/',
-    avatar: {
-      url: 'https://adamanderssongolf.com/wp-content/uploads/sb-instagram-feed-images/517973384_18507897097053711_8842400633805056808_nfull.webp',
-      width: 640,
-      height: 843,
-    },
-  },
-];
-
-export default function StartPage({
-  pastCompetitions,
-  upcomingCompetitions,
-  nextCompetition,
-  currentCompetition,
-  now: nowMs,
-}) {
-  pastCompetitions.forEach(ensureDates);
-  upcomingCompetitions.forEach(ensureDates);
-  if (nextCompetition) {
-    ensureDates(nextCompetition);
-  }
-  if (currentCompetition) {
-    ensureDates(currentCompetition);
-  }
-  const now = new Date(nowMs);
-
-  return (
-    <div className="chrome">
-      <Head>
-        <meta
-          name="description"
-          content={`
-${process.env.NEXT_PUBLIC_INTRO}. Follow your favorite players and get the latest updates straight in your inbox.
-        `.trim()}
-        />
-      </Head>
-      <Menu />
-      <div className="competitions">
-        <h2>{process.env.NEXT_PUBLIC_INTRO_TITLE}</h2>
-        <p className="page-desc">
-          {process.env.NEXT_PUBLIC_INTRO} Follow your{' '}
-          <Link href="/players">favorite players</Link>{' '}
-          and get the latest updates{' '}
-          <Link href="/profile">straight in your inbox</Link>
-          .
-        </p>
-        {currentCompetition && (
-          <Leaderboard competition={currentCompetition} now={now} />
-        )}
-        {nextCompetition ? (
-          <>
-            <h3>Next event</h3>
-            <ul>
-              <CompetitionListItem competition={nextCompetition} now={now} />
-            </ul>
-          </>
-        ) : null}
-        {upcomingCompetitions.length > 0 && (
-          <>
-            <h3>Future events</h3>
-            <ul>
-              {upcomingCompetitions.map(c => (
-                <CompetitionListItem key={c.id} competition={c} now={now} />
-              ))}
-            </ul>
-            <Link href="/schedule" className="page-margin competition-view-all">
-              View all events
-            </Link>
-          </>
-        )}
-
-        <h3 style={{ marginBottom: 0 }}>More from the players</h3>
-        <p className="page-desc">
-          Support players by following them on their social media accounts,
-          subscribing to their channels and watching their videos.
-        </p>
-        <ul className="youtubers">
-          {youtubers.map(y => (
-            <li key={y.name} className="youtuber">
-              <a
-                href={y.url}
-                className="youtuber-link"
-                target="_blank"
-                rel="noopener noreferrer"
-              >
-                <img
-                  src={y.avatar.url}
-                  alt={y.name}
-                  width={y.avatar.width}
-                  height={y.avatar.height}
-                />
-                {y.name}
-              </a>
-            </li>
-          ))}
-        </ul>
-
-        {pastCompetitions.length > 0 && (
-          <>
-            <h3>Past events</h3>
-            <ul>
-              {pastCompetitions.map(c => (
-                <CompetitionListItem key={c.id} competition={c} now={now} />
-              ))}
-            </ul>
-            <Link href="/schedule" className="page-margin competition-view-all">
-              View all events
-            </Link>
-          </>
-        )}
-      </div>
-    </div>
-  );
-}
-
-function CompetitionListItem({ competition, now, current }) {
-  const queryString = now > competition.end ? '?finished=1' : '';
-  return (
-    <li
-      key={competition.id}
-      className={
-        current ? 'competition-list-item current' : 'competition-list-item'
-      }
-    >
-      <Link href={`/t/${competition.slug}${queryString}`} className="competition">
-        <div className="calendar-event">
-          <b>{format(competition.start, 'd')}</b>
-          <span>{format(competition.start, 'MMM')}</span>
-        </div>
-        <div className="competition-details">
-          <h4 className="competition-name">
-            <span>{competition.name}</span>
-          </h4>
-          <p>
-            {competition.venue} — {competitionDateString(competition, now)}
-          </p>
-        </div>
-      </Link>
-    </li>
-  );
-}
+export default StartPage;
 
 export async function getServerSideProps() {
   const competitions = await prisma.competition.findMany({
